Добавление газовой сети и топологической сортировки
- Добавлена функциональность для соединения двух компрессорных станций с трубой.
- Проверка введенных ID станций на неотрицательность и уникальность, чтобы они не совпадали.
- Проверка существующих соединений, чтобы избежать дублирования.
- Поиск уже имеющихся труб заданного диаметра; если такие отсутствуют, пользователь может создать новую трубу.
- Обработаны условия для ситуаций, когда подходящая труба найдена или нет, с соответствующими сообщениями об ошибках и действиями.
- Улучшено введение пользователем значений диаметра труб.
- Отображение деталей соединения после каждой операции.
- Добавлена функция удаления соединения по его ID.
- Добавлен метод dfs для поиска в глубину с проверкой на наличие циклов.
- Реализован метод findConnectedComponents для нахождения связанных компонентов графа и их обработки.
- Добавлен метод topologicalSort для выполнения топологической сортировки на основе найденных компонентов.
- Оптимизирована работа с графом: реализована валидация на наличие циклов, что предотвращает выполнение топологической сортировки для циклических графов.
- При успешной сортировке выводится топологический порядок и матрица смежности с весами.

+++ Lapin_lab2/Pipe.h
@@ -12,8 +12,8 @@ class PipeManager;
 class Pipe {
 private:
     string full_name; 
-    double length = 0; 
-    double diameter = 0;
+    int length = 0; 
+    int diameter = 0;
     bool underRepair;
 public:
     void read();
